medical-paper-alerts/requirements.txt
# 医学論文自動通知システム - 必要なパッケージ

# PubMed API
biopython>=1.81
metapub>=0.6.4

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# YAML configuration
pyyaml>=6.0

# Claude API (Anthropic)
anthropic>=0.18.0

# Email
secure-smtplib>=0.1.1

# Scheduling
schedule>=1.2.0
python-crontab>=3.0.0

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# Logging
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
